requirements.txt
combo==0.1.2
Cython==0.29.21
cycler==0.10.0
joblib==1.0.0
jupyterlab==3.0.5
kiwisolver==1.3.1
llvmlite==0.35.0
matplotlib==3.3.3
networkx==2.5
numba==0.52.0
numpy==1.19.5
pandas==1.1.5
patsy==0.5.1
Pillow==8.1.0
psutil==5.8.0
pyod==0.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
spartan2==0.1.1.post0
statsmodels==0.12.1
suod==0.0.5
threadpoolctl==2.1.0
umap-learn==0.5.0
